refactor: controller 패키지명을 purchasable로 수정
